userprofile依赖models的函数从utils移动到services
